Fixing how histograms are presented to Librato
For Librato, count must be the number of measurements in this sample. It will show sum/count as the mean.
Sample.Size() gives us this. Sample.Count() gives the total number of measurements ever recorded for the
life of the histogram, which means the Librato graph will trend toward 0 as more measurements are recored.